requirements.txt
xue
pytest
uvicorn
fastapi
aiofiles
greenlet
aiosqlite
sqlalchemy
watchfiles
ruamel.yaml
httpx[http2]
cryptography
python-multipart